fix type

+++ pennylane/optimize/adaptive.py
@@ -53,7 +53,7 @@ class AdaptiveOptimizer:
     algorithms such as `ADAPT-VQE <https://www.nature.com/articles/s41467-019-10988-2>`_.
 
     Args:
-        param_steps (float): number of steps for optimizing the parameter of a selected gate
+        param_steps (int): number of steps for optimizing the parameter of a selected gate
         stepsize (float): step size for optimizing the parameter of a selected gate
 
     **Example**
